Refactor: create separate folder for services
