+++ src/components/card/Card.css
@@ -2,4 +2,5 @@
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
+    cursor: pointer;
 }
