+++ src/components/card/CardTextContent.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import { Link } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { useCart, useAuth } from "../../context";
-import { changeCartItemQtyService } from "./services/changeCartItemQty.service";
+import { changeCartItemQtyService } from "../../services";
 
 export function CardTextContent({
   title = "Velfit's Dumbbell",
@@ -11,6 +11,7 @@ export function CardTextContent({
   ratingValue = 1,
   product,
 }) {
+  const navigate = useNavigate();
   const [loader, setLoader] = useState(false);
   const {
     authState: { token },
@@ -18,6 +19,7 @@ export function CardTextContent({
   const { cartDispatch } = useCart();
 
   const handleIncCartItemClick = (e) => {
+    e.stopPropagation();
     setLoader(true);
     if (quantity > 0) {
       const requestObj = {
@@ -31,15 +33,18 @@ export function CardTextContent({
     }
   };
 
+  const navigateToProduct = () => {
+    navigate(`/products/${product._id}`);
+  };
+
   return (
-    <Link to={`/products/${product._id}`}>
-    <div className="card-text-content m-xs">
+    <div onClick={navigateToProduct} className="card-text-content m-xs">
       <p className="card-title"> {title} </p>
       <p className="subtitle"> {description} </p>
       <p className="price"> ₹ {price} </p>
       <div className="ecom-qty">
         <button
-          onClick={handleIncCartItemClick}
+          onClick={(e) =>  handleIncCartItemClick(e)}
           name="decrement"
           className="qty-btn fs-3 p-xxs dec"
           disabled={quantity < 2 || loader}
@@ -48,7 +53,7 @@ export function CardTextContent({
         </button>
         <span className="qty-num p-xxs"> {quantity} </span>
         <button
-          onClick={handleIncCartItemClick}
+          onClick={(e) =>  handleIncCartItemClick(e)}
           className="qty-btn fs-3 p-xxs inc"
           name="increment"
           disabled={loader}
@@ -65,6 +70,5 @@ export function CardTextContent({
         <span className="rating-info"> {ratingValue} </span>
       </div>
     </div>
-    </Link>
   );
 }
