--- src/components/card/services/index.js
@@ -1,5 +1,0 @@
-export { addToWishlistService } from "./addToWishlist.service";
-export { addToCartService } from "./addToCart.service";
-export { removeFromCartService } from "./removeFromCart.service";
-export { changeCartItemQtyService } from "./changeCartItemQty.service";
-export { removeFromWishlistService } from "./removeFromWishlist.service";
