+++ src/services/index.js
@@ -1,3 +1,9 @@
 export { addToCartService } from "./cartServices/addToCart.service";
 export { addToWishlistService } from "./wishlistServices/addToWishlist.service";
 export { removeFromWishlistService } from "./wishlistServices/removeFromWishlist.service";
+export { removeFromCartService } from "./cartServices/removeFromCart.service";
+export { changeCartItemQtyService } from "./cartServices/changeCartItemQty.service";
+export { loginService } from "./authServices/login.service";
+export { signupService } from "./authServices/signup.service";
+export { getCategoriesService } from "./getDataServices/getCategories.service";
+export { getProductService } from "./getDataServices/getProduct.service";
